feat: deletedCacheData 함수에서 예외 처리 추가
- 확장자가 없을 시 error 발생
- 삭제 성공 시 삭제한 요소 id 반환
- 삭제할 파일이 존재하지 않을 시 메세지 반환
- 실패 시 에러 메세지 반환

+++ src/shared/cache/json.ts
@@ -32,12 +32,19 @@ export const getCacheData = (src: string) => {
 };
 
 export const deletedCacheData = (src: string) => {
-  const filePath = path.join(process.cwd(), src);
-  if (fs.existsSync(filePath)) {
-    const fileContent = fs.rmdirSync(filePath);
-    return { message: "success" };
+  try {
+    const extension = src.split(".").at(-1);
+    if (!extension) return new Error("파일 확장자를 작성하지 않았습니다.");
+    const filePath = path.join(process.cwd(), src);
+    if (fs.existsSync(filePath)) {
+      fs.rmdirSync(filePath);
+      return { message: "success", id: src };
+    }
+    return { message: "파일이 존재하지 않습니다.", id: src };
+  } catch (error) {
+    console.error("deletedCacheData: 파일을 삭제하는 데 실패했습니다.");
+    throw new Error(`${error}`);
   }
-  return null;
 };
 
 export const updateJSONFile = <T>(
